backend/requirements.txt
fastapi==0.95.2
uvicorn==0.22.0
sqlalchemy==1.4.48
mysql-connector-python==8.0.33
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.6
pydantic==1.10.9
python-dotenv==1.0.0
email-validator==2.0.0
bcrypt==4.1.2